pyproject.toml
[project]
name = "hi-nelow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "numpy>=2.2.6",
    "openai>=1.81.0",
    "openai-whisper>=20240930",
    "playwright>=1.52.0",
    "pyaudio>=0.2.14",
]
